requirements.txt
django==2.1.2
django-filter==2.0.0
django-bootstrap3==11.0.0
django-widget-tweaks==1.4.3
elasticsearch==5.5.3
django-haystack-elasticsearch==0.1.0
psycopg2-binary==2.7.5
